+++ cucumber-expressions/ruby/spec/cucumber/cucumber_expressions/tree_regexp_spec.rb
@@ -54,11 +54,17 @@ module Cucumber
       end
 
       it 'captures non capturing groups with capturing groups inside' do
-        tr = TreeRegexp.new("the stdout(?: from \"(.*?)\")?")
+        tr = TreeRegexp.new(/the stdout(?: from "(.*?)")?/)
         group = tr.match("the stdout")
         expect(group.value).to eq("the stdout")
         expect(group.children[0].value).to eq(nil)
         expect(group.children.length).to eq(1)
+      end
+
+      it 'works with flags' do
+        tr = TreeRegexp.new(/HELLO/i)
+        group = tr.match("hello")
+        expect(group.value).to eq("hello")
       end
     end
   end